fix: fatal error handling

+++ lib/lpc/types.h
@@ -25,7 +25,7 @@ typedef struct {
 union svalue_u {
     char *string;
     const char *const_string;
-    int number;     /* Neolith extension: int is as wide as pointers */
+    long number;    /* Neolith extension: int is as wide as pointers */
     double real;    /* Neolith extension: both float and double are in native double precision */
 
     refed_t *refed; /* any of the block below */
